Reporting-Tool/requirements.txt
functions-framework==3.*
paho-mqtt
google.cloud.storage
pandas